+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ---
 
+## [Unreleased]
+
+### Added
+
+* Information about the error codes in the readme.md file.
+
 ## [**1.3.0**] - 2019-12-19
 
 ### Added